ELEC-316: Update current sense driver to work with isoSPI (#158)
The LTC6820 requires a clock signal to encode a change in state on the
MISO pin. Since the LTC2484 uses the MISO pin to signal End of
Conversion, we are unable to test that correctly, and thus we are forced
to delay for the maximum conversion time before performing another read.

In addition, this fixes the ripple test and changes the driver to use a
callback-driven model.