ELEC-105: Chaos Main (#216)
Finally after many PRs Chaos is in what I consider to be a "working" state. (At least all the modules independently are). Many tests are still required to validate its integration.

This provides the main for Chaos which is pretty simple.

NOTE: Thermistor controls and fan speed code is not present as this is being worked on presently and shared between multiple boards. In the interim I can just make the fans run full speed via PWM if we want.