ELEC-439: LTC AFE FSM (#230)
Switches the LTC6804 module from blocking to event-driven. This is necessary because the required delays block CAN processing, resulting in ACK timeouts.

The implementation layer splits conversion triggering from data readback for the FSM layer to delay and sequence appropriately. The intended usage is to register callbacks for the cell/aux conversion results.