ELEC-500: LTC AFE FSM Faults (#290)
Implements wakeup method 2 as specified in the datasheet. This makes wakeup much more reliable and actually allows us to reduce our delay between triggering and reading to much more reasonable values.

I plan on trying to bring the delay as low as possible, and if it seems to be stable on the order of a few milliseconds, we can probably switch back to a main loop polled method.