ELEC-534: Add CAN dump parser script (#303)
This adds a parse_data.py script that parses the logged data.

I've also added a README for the can_dump project's scripts.

In addition, I've refactored out the CAN parsing logic out of the main
dump_system.py script, and created a can_message.py file that
contains all that logic. This allows the parse_data.py script to share
the same code as the dump_system script.