ELEC-538: Add FreeRTOS libraries and example project (#308)
This adds FreeRTOS 10.1.1 code for the ARM Cortex M0 and M3 cores as 
a static library, as well as an implementation of the example project. 
The idea is that projects will not really need specific FreeRTOS 
configuration and can simply include FreeRTOS in their rules.mk file. 
The code is taken as-is from the FreeRTOS project webpage, with some 
minor changes to suppress some warnings when compiling on x86:

  - FreeRTOS/tasks.c: suppress 'pointer-to-int-cast'
  - FreeRTOS/tasks.c: suppress 'int-to-pointer-cast'

FreeRTOS is currently set up to only use static allocation, hence the
removal of the heap allocators. A shared Idle Task is provided in the
library, as we do not expect the Idle Task to change between projects.

This change also provides a stubbed out x86 implementation of FreeRTOS 
such that it can compile on x86. Currently, this does not do anything, 
and attempting to run the x86 build will probably not work as expected.

For some reason the temporary breakpoints in the gdb_flash script
appear to cause an assertion failure when debugging FreeRTOS. As we 
currently still need to support non-FreeRTOS projects, a quick 
workaround is to simply comment out the following lines when debugging 
a FreeRTOS project and set the first breakpoint after the scheduler:

  tb main
  tb LoopForever

In addition, LTO has been disabled for now, since we've been having 
issues with LTO and our current GCC version (GCC ARM 6-2017-q2-up).