ELEC-583: Fix LTC AFE discharge indices (#318)
This fixes a bug in the AFE driver that was introduced during some
refactoring in the cell discharge index mappings. This resulted in
the device and cell values used for the reverse discharge mapping
lookup being flipped.

As such, you would set the wrong index, due to reading the wrong
buffer. In fact, some cases would result in reads past the buffer's
memory bounds, resulting in checking and setting the incorrect index.

An additional LTC AFE test has been added that tests the mapping.