ELEC-588: Remove reallyclean target (#321)
Remove the distinction between the clean and reallyclean targets.

This creates a lot of confusion, and there are very few cases where the
intended behaviour is to only clean a specific project. Moreover, in the
interests of keeping things consistent with typical conventions, this
renames the reallyclean target to clean.