WIP: Revert changes in ADC driver
Some regressions were introduced to the ADC driver when the Thermistor
Reading changes were added. In particular, this broke the ADC Driver
test project, and presumable other use cases with the introduction of a
volatile variable that was loaded/stored when checking if conversions
where complete, instead of using the StdPeriph drivers that were
originally being used.