ELEC-580: Add Pedal FSMs and project skeleton
Add the Pedal FSMs that are not dependant on Drive Output. These FSMs
are only dependent on input from Driver Controls: Center Console and
Driver Controls: Steering.

An Event Arbiter is introduced in order to block FSMs from accepting
input in certain states. Technically, this could be flattend into a
giant FSM by computing the Transitive Closure, but it probably makes it
a little easier to reason about and modify this way. This implementation
is pretty much taken from the old Driver Controls code.