ELEC-636: Remove redundant linter case (#343)
The whitespace/braces case is pretty much handled implicitly by our
clang-format rules. In certain cases, this clashes with the formatting
that clang-format provides, resulting in no way to pass CI.

As a fix, we remove this check from our cpplint.py fork, as we check
that clang-format has been applied on each PR anyways.