ELEC-580: Add Pedal FSMs and project skeleton (#342)
Add the Pedal FSMs that are not dependent on Drive Output. These FSMs
are only dependent on input from Driver Controls: Center Console and
Driver Controls: Steering.

An Event Arbiter is introduced in order to block FSMs from accepting
input in certain states. Technically, this could be flattened into a
giant FSM by computing the Transitive Closure, but it probably makes it
a little easier to reason about and modify this way.