ELEC-580: Add Center Console button handling
Add logic to handle GPIO inputs for the Center Console buttons and raise
the appropriate events over CAN.

Due to a HW bug (see HW-200), we can't handle all the inputs via a GPIO
interrupt. As such, we poll for the input that shares an EXTI line in
the main loop.