ELEC-580: Add Center Console BPS Fault Indicator
Add code to drive the BPS Fault Dash Indicator LED by piggybacking the
CAN message used to notify Lights to activate the Strobe. This is
because there's currently some inherent coupling between the Lights
board, which is told when to turn on the Strobe, and the required BPS
Fault Dash Indicator.

If we wish to process the message slightly faster instead of waiting for
the Driver Controls master board to notify everyone, we could also
listen to the BMS Heartbeat.