ELEC-580: Add Pedal dependencies for drive output
Add the FSMs and logic that controls the Drive Output message.

This includes a basic way of toggling whether regenerative braking is
enabled via a compile-time constant. This should be further extended in
the future (either at the Motor Controller Interface or in the Driver
Controls Master) to either allow the driver to configure whether regen
is enabled, or dynamically do so depending on battery SoC.

Currently it also naively assumes that the ADS1015 is always going to
successfully sample. It is probably worth adding logic to the FSMs to
deal with the failure case where the ADC stops sampling, resulting in
stale data being sent to the Motor Controller Interface. Currently we
just zero the data for the pedal/brake positions and hope for the best.

In addition, modified the Exported Enums to add the bitmasks for BMS
Heartbeat status, since it seems like the enum indices were updated, but
this was missed in the process.