ELEC-580: Remove old Driver Controls code
This project is no longer used in MS XII.

The firmware was originally used in MS XII prior to FSGP/ASC 2018, but
due to various issues around EMC, the hardware was heavily modified for
competition. It was later superseded by the distributed Driver Controls
hardware, consisting of the Pedal, Steering, and Center Console boards,
which were used at FSGP 2019.